requirements/requirements_3-7.txt
appdirs==1.4.3
asyncpg==0.18.3
atomicwrites==1.3.0
attrs==19.1.0
black==19.3b0
boto3==1.9.224
botocore==1.12.224
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.4
dj-database-url==0.5.0
django-cors-headers==1.2.2
django-debug-toolbar==1.7
django-extensions==1.7.5
django-filter==0.14.0
django-partial-index==0.5.2
django-queryset-csv==1.0.1
django-simple-history==1.8.2
django-spaghetti-and-meatballs==0.2.2
Django>=1.11.0,<2.0
djangorestframework==3.9.4
docker==4.1.0
docutils==0.15.2
dredd-hooks==0.2.0
drf-extensions==0.3.1
drf-tracking==1.4.0
elasticsearch==6.3.0
entrypoints==0.3
et-xmlfile==1.0.1
filechunkio==1.8
fiscalyear==0.1.0
flake8==3.7.8
git+https://github.com/fedspendingtransparency/django-mock-queries#egg=django-mock-queries
idna==2.8
importlib-metadata==0.20
jdcal==1.4.1
jmespath==0.9.4
jsonpickle==1.2
Markdown==3.1.1
mccabe==0.6.1
mock==3.0.5
model-mommy==1.6.0
more-itertools==7.2.0
numpy==1.17.2
openpyxl==2.4.7
pandas==0.25.1
pip>=19.2
pluggy==0.12.0
psycopg2-binary==2.7.5
py-gfm==0.1.3
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytest-cov==2.8.1
pytest-django==3.5.1
pytest==4.3.1
python-dateutil==2.7.3
python-json-logger==0.1.9
pytz==2019.2
requests==2.22.0
retrying==1.3.3
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
unicodecsv==0.14.1
urllib3>=1.25,<1.26
xlrd==1.0.0
zipp==0.6.0
